restructure history recording

+++ server/endpoints/userhistoryrecord.js
@@ -1,14 +1,15 @@
 const { PrismaClient } = require('@prisma/client');
 const prisma = new PrismaClient();
 
-module.exports.name = "/api/user/history/toggle/:toonslug/:chapterorder";
+module.exports.name = "/api/user/history/record/:toonslug/:chapterorder";
 module.exports.method = "POST";
 module.exports.verify = function (req, res) {
     return req.auth;
 }
 
 module.exports.execute = function (req, res) {
-    if(req.params.toonslug && req.params.chapterorder && !isNaN(parseInt(req.params.chapterorder))) {
+    let todelete = !!req.query.delete;
+    if (req.params.toonslug && req.params.chapterorder && !isNaN(parseInt(req.params.chapterorder))) {
         prisma.chapter.findFirst({
             where: {
                 toon: {
@@ -17,14 +18,14 @@ module.exports.execute = function (req, res) {
                 order: parseInt(req.params.chapterorder)
             }
         }).then((chapter) => {
-            if(chapter) {
+            if (chapter) {
                 prisma.readHistory.findFirst({
                     where: {
                         userid: req.auth.id,
                         chapterid: chapter.id
                     }
                 }).then((history) => {
-                    if(history) {
+                    if (history && todelete) {
                         prisma.readHistory.delete({
                             where: {
                                 userid: req.auth.id,
@@ -34,7 +35,7 @@ module.exports.execute = function (req, res) {
                             res.status(200).json({ status: 200, message: `Read history deleted` });
                         });
                     }
-                    else {
+                    else if (!history && !todelete) {
                         prisma.readHistory.create({
                             data: {
                                 userid: req.auth.id,
@@ -43,6 +44,9 @@ module.exports.execute = function (req, res) {
                         }).then(() => {
                             res.status(201).json({ status: 201, message: `Read history created` });
                         });
+                    }
+                    else {
+                        res.status(204).json({ status: 204, message: `No change` });
                     }
                 });
             }
